+++ boot.py
@@ -196,5 +196,6 @@ print( "call loop() to start :-)" )
 print()
 
 
+#from samples.windup_fiber import serve
 
 #end